requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.49
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6
pydantic==2.13.3
pydantic-settings==2.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
tenacity==9.1.4
anthropic>=0.45.0
openai>=1.0.0
google-genai>=1.0.0
langchain-core>=0.3.34,<1.0.0
langchain-community>=0.3.10,<1.0.0
langchain-anthropic>=0.3.7
langchain-groq>=0.2.0
langgraph>=0.2.68,<1.0.0
numpy==2.4.4
scikit-learn==1.6.1
joblib==1.4.2
pandas==2.2.3
sentence-transformers==3.4.0
structlog==24.4.0
pytest==8.3.4
pytest-asyncio==0.25.2
jupyter==1.1.1
matplotlib==3.10.0
seaborn==0.13.2